move all parsing functions to the parsing module
